pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atts"
version = "2.0.0"
description = "Adaptive Test-Time Scaling - Clean Architecture with Explainable AI"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "ATTS Contributors"}
]
dependencies = [
    "ollama>=0.4.0",
    "PyYAML>=6.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "numpy>=1.24.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
atts = "atts.interfaces.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
